:recycle: [Refactor] 공통 컴포넌트 및 스타일 적용
