--- src/page/driver/login/components/PageInput.tsx
@@ -1,99 +1,0 @@
-import { useState } from "react";
-import style from "./PageInput.module.css";
-
-type InputType = {
-  value: string;
-  inputHeandler: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  validation: boolean;
-};
-
-export const EmailInputComponent = ({
-  value,
-  inputHeandler,
-  validation,
-}: InputType) => {
-  return (
-    <div className={style.container}>
-      <span className={style.text}>이메일</span>
-      <input
-        className={`${style.input} ${
-          style[value && validation ? "complete" : ""]
-        } ${style[value && !validation ? "invalid" : ""]} `}
-        value={value ?? ""}
-        onChange={inputHeandler}
-        type="text"
-        id="email"
-        name="email"
-        placeholder="이메일을 입력 해주세요."
-      />
-      {validation ? null : (
-        <span className={style.invaild}>이메일 형식이 아닙니다.</span>
-      )}
-    </div>
-  );
-};
-
-export const PasswordInputComponent = ({
-  value,
-  inputHeandler,
-  validation,
-}: InputType) => {
-  const [invisible, setInvisible] = useState<Boolean>(true);
-
-  return (
-    <div className={style.container}>
-      <span className={style.text}>비밀번호</span>
-      <div className={style.inputContainer}>
-        <input
-          value={value ?? ""}
-          onChange={inputHeandler}
-          className={`${style.input} ${
-            style[value && validation ? "complete" : ""]
-          } ${style[value && !validation ? "invalid" : ""]} `}
-          type={invisible ? "password" : "text"}
-          id="password"
-          name="password"
-          placeholder="비밀번호을 입력 해주세요."
-        />
-        <div className={style.iconPill}>
-          {invisible ? (
-            <svg
-              onClick={() => {
-                setInvisible((prev) => !prev);
-              }}
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M15.7726 12.9729L14.6496 11.8499C14.8201 10.9601 14.5861 10.2012 13.9477 9.57296C13.3092 8.94476 12.5432 8.70374 11.6496 8.84989L10.5265 7.72684C10.7586 7.61914 10.9974 7.54157 11.2429 7.49414C11.4884 7.44671 11.7406 7.42299 11.9996 7.42299C13.1342 7.42299 14.0973 7.81882 14.889 8.61049C15.6807 9.40216 16.0765 10.3653 16.0765 11.4999C16.0765 11.7589 16.0544 12.0111 16.0101 12.2566C15.9659 12.5021 15.8867 12.7409 15.7726 12.9729ZM18.9534 16.0845L17.8496 15.0499C18.4829 14.5666 19.0496 14.0291 19.5496 13.4374C20.0496 12.8457 20.4663 12.1999 20.7996 11.4999C19.9663 9.81656 18.7621 8.47906 17.1871 7.48739C15.6121 6.49572 13.8829 5.99989 11.9996 5.99989C11.5163 5.99989 11.0454 6.03322 10.5871 6.09989C10.1288 6.16656 9.66625 6.26656 9.19959 6.39989L8.03424 5.23454C8.67012 4.98326 9.31851 4.798 9.97941 4.67877C10.6403 4.55953 11.3137 4.49991 11.9996 4.49991C14.1483 4.49991 16.122 5.05825 17.9207 6.17491C19.7194 7.29158 21.1028 8.80247 22.0707 10.7076C22.1374 10.8345 22.1858 10.9624 22.2159 11.0912C22.246 11.2201 22.2611 11.3563 22.2611 11.4999C22.2611 11.6435 22.2486 11.7797 22.2236 11.9085C22.1986 12.0374 22.1528 12.1653 22.0861 12.2922C21.7322 13.0448 21.288 13.7396 20.7534 14.3768C20.2188 15.014 19.6188 15.5832 18.9534 16.0845ZM11.9996 18.4999C9.89447 18.4999 7.96499 17.9374 6.21116 16.8124C4.45731 15.6874 3.06308 14.2056 2.02846 12.3672C1.94513 12.2403 1.88424 12.1025 1.84579 11.9537C1.80732 11.805 1.78809 11.6537 1.78809 11.4999C1.78809 11.346 1.80475 11.1973 1.83809 11.0537C1.87142 10.9102 1.92975 10.7698 2.01309 10.6326C2.38489 9.95311 2.80604 9.30599 3.27656 8.69124C3.74708 8.07649 4.2881 7.52553 4.89964 7.03837L2.64194 4.76526C2.50347 4.61655 2.4352 4.43995 2.43714 4.23546C2.43905 4.03098 2.51244 3.85631 2.65731 3.71144C2.80218 3.56657 2.97782 3.49414 3.18424 3.49414C3.39064 3.49414 3.56627 3.56657 3.71114 3.71144L20.288 20.2883C20.4265 20.4268 20.4999 20.5983 20.5082 20.8027C20.5165 21.0072 20.4432 21.187 20.288 21.3422C20.1432 21.487 19.9675 21.5595 19.7611 21.5595C19.5547 21.5595 19.3791 21.487 19.2342 21.3422L15.715 17.8537C15.1252 18.0819 14.5191 18.2467 13.8967 18.3479C13.2743 18.4492 12.6419 18.4999 11.9996 18.4999ZM5.95346 8.09216C5.36756 8.54473 4.84095 9.0566 4.37364 9.62777C3.90632 10.1989 3.51497 10.823 3.19959 11.4999C4.03292 13.1832 5.23709 14.5207 6.81209 15.5124C8.38709 16.5041 10.1163 16.9999 11.9996 16.9999C12.4291 16.9999 12.8519 16.971 13.2679 16.9133C13.6839 16.8557 14.0971 16.7666 14.5073 16.646L13.2419 15.3499C13.0393 15.4383 12.8371 15.4982 12.6352 15.5297C12.4332 15.5611 12.2214 15.5768 11.9996 15.5768C10.865 15.5768 9.90185 15.181 9.11019 14.3893C8.31852 13.5976 7.92269 12.6345 7.92269 11.4999C7.92269 11.2781 7.93999 11.0662 7.97461 10.8643C8.00923 10.6624 8.06756 10.4601 8.14961 10.2576L5.95346 8.09216Z"
-                fill="#ABABAB"
-              />
-            </svg>
-          ) : (
-            <svg
-              onClick={() => {
-                setInvisible((prev) => !prev);
-              }}
-              width="24"
-              height="24"
-              viewBox="0 0 24 24"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M12.0021 15.5769C13.1352 15.5769 14.0976 15.1803 14.8892 14.3871C15.6809 13.5939 16.0767 12.6308 16.0767 11.4977C16.0767 10.3646 15.6802 9.40224 14.887 8.61058C14.0938 7.81891 13.1307 7.42308 11.9976 7.42308C10.8645 7.42308 9.9021 7.81966 9.11044 8.61282C8.31877 9.40601 7.92294 10.3692 7.92294 11.5023C7.92294 12.6353 8.31952 13.5977 9.11269 14.3894C9.90587 15.181 10.869 15.5769 12.0021 15.5769ZM11.9998 14.2C11.2498 14.2 10.6123 13.9375 10.0873 13.4125C9.56234 12.8875 9.29984 12.25 9.29984 11.5C9.29984 10.75 9.56234 10.1125 10.0873 9.58748C10.6123 9.06248 11.2498 8.79998 11.9998 8.79998C12.7498 8.79998 13.3873 9.06248 13.9123 9.58748C14.4373 10.1125 14.6998 10.75 14.6998 11.5C14.6998 12.25 14.4373 12.8875 13.9123 13.4125C13.3873 13.9375 12.7498 14.2 11.9998 14.2ZM11.9998 18.5C9.89472 18.5 7.96941 17.9384 6.22391 16.8153C4.47839 15.6923 3.08513 14.2147 2.04411 12.3827C1.96078 12.2391 1.89988 12.0944 1.86141 11.9485C1.82294 11.8027 1.80371 11.653 1.80371 11.4995C1.80371 11.346 1.82294 11.1965 1.86141 11.051C1.89988 10.9054 1.96078 10.7609 2.04411 10.6173C3.08513 8.78525 4.47839 7.30769 6.22391 6.18463C7.96941 5.06154 9.89472 4.5 11.9998 4.5C14.105 4.5 16.0303 5.06154 17.7758 6.18463C19.5213 7.30769 20.9145 8.78525 21.9556 10.6173C22.0389 10.7609 22.0998 10.9056 22.1383 11.0514C22.1767 11.1973 22.196 11.347 22.196 11.5005C22.196 11.654 22.1767 11.8035 22.1383 11.949C22.0998 12.0945 22.0389 12.2391 21.9556 12.3827C20.9145 14.2147 19.5213 15.6923 17.7758 16.8153C16.0303 17.9384 14.105 18.5 11.9998 18.5ZM11.9998 17C13.8832 17 15.6123 16.5041 17.1873 15.5125C18.7623 14.5208 19.9665 13.1833 20.7998 11.5C19.9665 9.81664 18.7623 8.47914 17.1873 7.48748C15.6123 6.49581 13.8832 5.99998 11.9998 5.99998C10.1165 5.99998 8.38734 6.49581 6.81234 7.48748C5.23734 8.47914 4.03317 9.81664 3.19984 11.5C4.03317 13.1833 5.23734 14.5208 6.81234 15.5125C8.38734 16.5041 10.1165 17 11.9998 17Z"
-                fill="#ABABAB"
-              />
-            </svg>
-          )}
-        </div>
-      </div>
-      {validation ? null : (
-        <span className={style.invaild}>비밀번호가 올바르지 않습니다.</span>
-      )}
-    </div>
-  );
-};
