--- src/page/user/login/components/LoginBtn.module.css
@@ -1,22 +1,0 @@
-.container {
-  display: flex;
-  flex-shrink: 0;
-  justify-content: center;
-  align-items: center;
-
-  width: 100%;
-  height: 64px;
-
-  border-radius: 16px;
-  background-color: #d2d2d2;
-
-  color: white;
-  font-size: 18px;
-  font-weight: 600;
-
-  cursor: pointer;
-}
-
-.container.complete {
-  background-color: #1b92ff;
-}
