--- src/page/user/login/components/LoginBtn.tsx
@@ -1,16 +1,0 @@
-import style from "./LoginBtn.module.css";
-
-type Props = {
-  context: string;
-  validation: boolean;
-};
-
-export default function LoginBtn({ context, validation }: Props) {
-  return (
-    <button
-      className={`${style.container} ${style[validation ? "complete" : ""]}`}
-    >
-      {context}
-    </button>
-  );
-}
