--- src/page/user/login/components/PageInput.module.css
@@ -1,81 +1,0 @@
-.none {
-  display: none;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: start;
-  justify-content: center;
-
-  width: 100%;
-  min-height: 112px;
-  max-height: 146px;
-  gap: 16px;
-}
-.text {
-  font-size: 20px;
-  font-weight: 400;
-  color: black;
-}
-.input {
-  width: 100%;
-  height: 64px;
-
-  box-sizing: border-box;
-  margin: 0;
-  padding-left: 14px;
-  padding-top: 16px;
-  padding-bottom: 16px;
-
-  border-radius: 16px;
-  border: 1px solid #e6e6e6;
-
-  font-size: 20px;
-  font-weight: 400;
-  color: black;
-}
-
-.input.complete {
-  border: 1px solid #1b92ff;
-}
-.input.invalid {
-  border: 1px solid #ff4f64;
-}
-.input:focus.invalid {
-  border: 1px solid #ff4f64;
-}
-
-.input:focus {
-  border: 1px solid #1b92ff;
-}
-
-.invaild {
-  display: flex;
-  justify-content: end;
-
-  width: 100%;
-  font-size: 16px;
-  font-weight: 500;
-  color: #ff4f64;
-}
-
-.inputContainer {
-  display: flex;
-  width: 100%;
-
-  position: relative;
-}
-
-.iconPill {
-  width: 24px;
-  height: 24px;
-  border-radius: 50%;
-
-  position: absolute;
-  left: 93%;
-  top: 20px;
-}
-.iconPill:hover {
-  background-color: lightgray;
-}
