--- src/page/user/login/index.module.css
@@ -1,99 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 100%;
-  height: 100%;
-}
-
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  gap: 72px;
-
-  width: 100%;
-  max-width: 640px;
-  height: 100%;
-  max-height: 852px;
-
-  background-color: white;
-}
-
-p {
-  font-size: 20px;
-  font-weight: 400;
-}
-
-a {
-  font-size: 20px;
-  font-weight: 600;
-  color: #1b92ff;
-}
-
-a:visited {
-  font-size: 20px;
-  font-weight: 600;
-  color: #1b92ff;
-}
-
-.top {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  width: inherit;
-  height: 140px;
-  gap: 18px;
-}
-
-.mid {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  width: inherit;
-  height: 432px;
-  gap: 56px;
-}
-
-.loginForm {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  width: 100%;
-  height: 100%;
-
-  gap: 32px;
-}
-
-.bottom {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  width: inherit;
-  height: 432px;
-  gap: 32px;
-}
-
-.bottom > span {
-  font-size: 20px;
-  font-weight: 400;
-  color: #525252;
-}
-
-.snsIcon {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  gap: 32px;
-}
